python/requirements.txt
requests
beautifulsoup4
redis
hiredis
progressbar2
websockets
aioredis
zstd
sanic
orjson
python-socketio
